+++ jgclark.Summaries/CHANGELOG.md
@@ -1,11 +1,14 @@
 # What's Changed in ⏱ Habits and Summaries plugin?
 (And see the full [README](https://github.com/NotePlan/plugins/tree/main/jgclark.Summaries).)
 
-## [0.19.4] - 2023-09-26 unreleased ???
+## [0.20.0] - 2023-10-???
 ### Added
-- refresh button to output of /periodStats where the time period is "<period> to date"
-- ??? make this work with all options
+- new **heatmap for tag** command that displays a 'heatmap' chart of a chosen tag's values for each day (e.g. all `@work(...)` values from daily notes)
+- ??? make below work with all options
 
+## [0.19.4] - 2023-09-26 unreleased
+### Added
+- refresh button to output of **periodStats** command, where the time period is "<period> to date"
 ## [0.19.3] - 2023-08-06
 ### Fixed
 - date logic on 'weeklyStatsToCSV' command output